Begin migrating old code to use long file id.

+++ clients/unshaded/src/main/java/tachyon/client/AbstractTachyonFS.java
@@ -120,7 +120,7 @@ abstract class AbstractTachyonFS implements TachyonFSCore {
    * @return true if succeed, false otherwise
    * @throws IOException
    */
-  public synchronized boolean rename(int fileId, TachyonURI dstPath) throws IOException {
+  public synchronized boolean rename(long fileId, TachyonURI dstPath) throws IOException {
     return rename(fileId, TachyonURI.EMPTY_URI, dstPath);
   }
